refactor(pr10): extract getSchemaField helper to @fields/ (G5i.1-G5i.5)
Creates src/fields/getSchemaField.mts with a single canonical implementation taking document + fieldPath. Replaces three inline copies (DocumentSheetStore, FieldOverridesStore, resolveChangeValue) with wrappers/calls that delegate to the shared helper.
