+++ src/fields/index.mts
@@ -5,6 +5,7 @@
  * the system live here. Use the `@fields/*` alias to import.
  */
 export * from './fieldBuilders.mjs';
+export { getSchemaField } from './getSchemaField.mjs';
 export { PriceData } from './PriceData.mjs';
 export { PriceField } from './PriceField.mjs';
 export type { SectionFieldOptions } from './SectionField.mjs';
